All used code during to produce thesis results
The Python file is the first file, as this was used to extract the text data from the SEC EDGAR database. Preprocessing.R is the second file, which merges the data. After phase 1, the files Talk Quantity.R and Seeded_LDA.R are employed before moving to phase 2 in Preprocessing.R where the variable Talk Quantity and the topics are added to the data. Lastly, Regression_models.R contains the panel regression, variable selection methods, and mediation model.